Split out CLI logs tests
They're a bit heavier than the basic tests and cover specific functionality.
